refactor: rename Models::Server
