+++ lib/textflight-client.rb
@@ -6,11 +6,11 @@ require "textflight-client/string_utils"
 require "textflight-client/dijkstra/dijkstra"
 require "textflight-client/command_parser"
 require "textflight-client/response_parser"
-require "textflight-client/models/model"
-require "textflight-client/models/nav"
-require "textflight-client/models/scan"
-require "textflight-client/models/status_report"
-require "textflight-client/models/status"
+require "textflight-client/models/server/base_models"
+require "textflight-client/models/server/nav"
+require "textflight-client/models/server/scan"
+require "textflight-client/models/server/status_report"
+require "textflight-client/models/server/status"
 require "textflight-client/models/local"
 require "textflight-client/tfprompt"
