+++ lib/textflight-client/response_parser.rb
@@ -2,122 +2,6 @@
 module TFClient
   class ResponseParser
 
-    FIELD_DELIMITER = "|".freeze
-    VARIABLE_REGEX = /(\{[a-z_]+\}+)/.freeze
-
-    def self.substitute_line_values(line:)
-      return line.chomp if !line[/\|/]
-      tokens = line.chomp.split("|")
-
-      translation = tokens[1]
-
-      matches = translation.scan(VARIABLE_REGEX)
-
-      return translation  if matches.empty?
-
-      values = self.hash_from_line_values(line: line.chomp)
-
-      with_substitutes = translation.chomp
-
-      matches.each do |match|
-        key = match[0].sub("{", "").sub("}", "").to_sym
-        with_substitutes.gsub!(match[0], values[key])
-      end
-
-      with_substitutes
-    end
-
-    def self.substitute_values(lines:)
-      lines.map do |line|
-        self.substitute_line_values(line:line.chomp)
-      end
-    end
-
-    def self.tokenize_line(line:)
-      lines = line.split(FIELD_DELIMITER)
-      stripped = []
-      lines.each_with_index do |line, index|
-        if index == 0
-          stripped << line
-        else
-          stripped << line.strip
-        end
-      end
-      stripped
-    end
-
-    # returns two values
-    def self.line_and_index_for_beginning_with(lines:, string:)
-      lines.each_with_index do |line, index|
-        return line.chomp, index if line.start_with?(string)
-      end
-      return nil, -1
-    end
-
-    # Returns a hash of the key=value pairs found at the end of lines
-    def self.hash_from_line_values(line:)
-      tokens = self.tokenize_line(line: line)[2..-1]
-      hash = {}
-      tokens.each do |token|
-        key_value = token.split("=")
-        hash[key_value[0].to_sym] = key_value[1]
-      end
-      hash
-    end
-
-    def self.is_list_item?(line:)
-      if line && line.length != 0 && line.start_with?("\t")
-        true
-      else
-        false
-      end
-    end
-
-    def self.collect_list_items(lines:, start_index:)
-      items = []
-      index = start_index
-      loop do
-        line = lines[index]
-        if self.is_list_item?(line: line)
-          items << line.strip
-          index = index + 1
-        else
-          break
-        end
-      end
-      items
-    end
-
-    def self.label_and_translation(tokens:)
-      if tokens[0][/Claimed by/]
-        {label: "Claimed by", translation: tokens[1].split("'")[0].strip}
-      else
-        {label: tokens[0].split(":")[0], translation: tokens[1].split(":")[0] }
-      end
-    end
-
-    def self.camel_case_from_string(string:)
-      string.split(" ").map do |token|
-        token.capitalize
-      end.join("")
-    end
-
-    def self.snake_case_sym_from_string(string:)
-      string.split(" ").map do |token|
-        token.downcase
-      end.join("_").to_sym
-    end
-
-    def self.model_class_from_string(string:)
-      if !TFClient::Models.constants.include?(string.to_sym)
-        return nil
-      end
-
-      "TFClient::Models::#{string}".split("::").reduce(Object) do |obj, cls|
-        obj.const_get(cls)
-      end
-    end
-
     attr_reader :command
     attr_reader :textflight_command
     attr_reader :response
@@ -139,17 +23,25 @@ module TFClient
       when "status"
         parse_status(command: @command)
       else
-        if @response[/#{Models::STATUS_BEGIN}/]
-          @response = @lines[0].chomp
-          @lines = [@response]
-        end
-
-        puts ResponseParser.substitute_values(lines: @lines).join("\n")
+        echo_response(command: @command)
       end
     end
 
+    def echo_response(command:)
+      if @response[/#{TFClient::Models::Server::STATUS_BEGIN}/]
+        TFClient.info("Received STATUS REPORT response")
+        return parse_status(command: command)
+      end
+
+      parser = Class.new do
+        include TFClient::Models::Server::Parser
+      end.new
+
+      puts parser.substitute_values(lines: @lines).join("\n")
+    end
+
     def parse_nav(command:)
-      nav = TFClient::Models::Nav.new(lines: lines)
+      nav = TFClient::Models::Server::Nav.new(lines: lines)
       if command != "nav-for-prompt"
         puts nav.response
       end
@@ -157,43 +49,19 @@ module TFClient
     end
 
     def parse_scan
-      scan = TFClient::Models::Scan.new(lines: lines)
+      scan = TFClient::Models::Server::Scan.new(lines: lines)
       puts scan.response
       scan
     end
 
     def parse_status(command:)
       if command == "status-for-prompt"
-        TFClient::Models::Status.new(lines: lines)
+        TFClient::Models::Server::Status.new(lines: lines)
       else
-        _, index_start =
-          ResponseParser.line_and_index_for_beginning_with(
-            lines: @lines,
-            string: Models::STATUS_BEGIN
-          )
-        if index_start == -1
-          puts ResponseParser.substitute_values(lines: @lines).join("\n")
-        end
-
-        _, index_end =
-          ResponseParser.line_and_index_for_beginning_with(
-            lines: @lines,
-            string: Models::STATUS_END
-          )
-
-        if index_start != 0
-          lines_before_status = @lines[0..index_start - 1]
-          puts ResponseParser.substitute_values(
-            lines: lines_before_status
-          ).join("\n")
-        else
-          lines_after_status = @lines[index_end + 1..-1]
-          puts ResponseParser.substitute_values(
-            lines: lines_after_status
-          ).join("\n")
-
-          Models::StatusReport.new(lines: @lines[index_start...index_end])
-        end
+        lines_to_print, status_report =
+          Models::Server::Parser.process_status_report_response(lines: lines)
+        puts lines_to_print.join("\n")
+        status_report
       end
     end
   end
